requirements.txt
streamlit
whisper
openai-whisper
deepl
google-generativeai
python-dotenv
werkzeug
